requirements.txt
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==2.2.16
djangorestframework==3.11.1
ipdb==0.13.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==2.0.10
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
win-unicode-console==0.5
